fix migration m_predikat_nilai

+++ database/migrations/1621021964609_m_predikat_nilai_schema.js
@@ -5,22 +5,25 @@ const Schema = use("Schema");
 
 class MPredikatNilaiSchema extends Schema {
   up() {
-    this.create("m_predikat_nilai", (table) => {
-      table.increments();
-      table.string("predikat");
-      table.integer("bb_pengetahuan");
-      table.integer("ba_pengetahuan");
-      table.integer("bb_keterampilan");
-      table.integer("ba_keterampilan");
-      table.string("sikap");
-      table.integer("m_sekolah_id").unsigned().index("m_sekolah_id");
-      table
-        .foreign("m_sekolah_id")
-        .references("m_sekolah.id")
-        .onDelete("cascade")
-        .onUpdate("cascade");
-      table.boolean("dihapus").defaultTo(0);
-      table.timestamps();
+    this.hasTable("m_predikat_nilai").then((exist) => {
+      if (!exist)
+        this.create("m_predikat_nilai", (table) => {
+          table.increments();
+          table.string("predikat");
+          table.integer("bb_pengetahuan");
+          table.integer("ba_pengetahuan");
+          table.integer("bb_keterampilan");
+          table.integer("ba_keterampilan");
+          table.string("sikap");
+          table.integer("m_sekolah_id").unsigned().index("m_sekolah_id");
+          table
+            .foreign("m_sekolah_id")
+            .references("m_sekolah.id")
+            .onDelete("cascade")
+            .onUpdate("cascade");
+          table.boolean("dihapus").defaultTo(0);
+          table.timestamps();
+        });
     });
   }
